+++ Commands/Logger.js
@@ -4,6 +4,7 @@
  */
 module.exports = {
     log: function(error) { //Logs error in .log File
-        console.log(error + '\n-------------------------------------------- \n\n');
-    },
+        console.log(error);
+        console.log('\n-------------------------------------------- \n\n');
+    }
 }